// src/app/API.d.ts
type IApiTrack = IAPIObject['data']['tracks'][0];
type ITrackExtend = {
    updated_artwork?: string;
    started?: boolean;
};

export type ICustomTrack = IApiTrack & ITrackExtend;

export type IAPI = IAPIObject & {
    data: {
        updated_artwork?: string;
    };
};